+++ internal/storage/node_session.go
@@ -77,7 +77,11 @@ func CreateNodeSession(p *redis.Pool, s models.NodeSession) error {
 	exp := int64(common.NodeSessionTTL) / int64(time.Millisecond)
 
 	if _, err := redis.String(c.Do("SET", fmt.Sprintf(nodeSessionKeyTempl, s.DevAddr), buf.Bytes(), "NX", "PX", exp)); err != nil {
-		return fmt.Errorf("create node-session %s for node %s error: %s", s.DevAddr, s.DevEUI, err)
+		if err == redis.ErrNil {
+			return fmt.Errorf("create node-session %s for node %s error: DevAddr already in use", s.DevAddr, s.DevEUI)
+		} else {
+			return fmt.Errorf("create node-session %s for node %s error: %s", s.DevAddr, s.DevEUI, err)
+		}
 	}
 	// DevEUI -> DevAddr pointer
 	if _, err := redis.String(c.Do("PSETEX", fmt.Sprintf(nodeSessionKeyTempl, s.DevEUI), exp, s.DevAddr.String())); err != nil {